Added new BldCommand trait for a minor refactoring of how cli sub commands are exposed and executed.
