--- src/config/cli.rs
@@ -1,17 +1,0 @@
-use crate::config::definitions::VERSION;
-use clap::{App, Arg, SubCommand};
-
-pub fn command() -> App<'static, 'static> {
-    let local = Arg::with_name("local")
-        .short("l")
-        .long("local")
-        .help("List configuration for local options");
-    let remote = Arg::with_name("remote")
-        .short("r")
-        .long("remote")
-        .help("List configuration for remote options");
-    SubCommand::with_name("config")
-        .about("Lists bld's configuration")
-        .version(VERSION)
-        .args(&[local, remote])
-}
