+++ src/main.rs
@@ -2,6 +2,7 @@ mod auth;
 mod config;
 mod helpers;
 mod hist;
+mod high_avail;
 mod init;
 mod inspect;
 mod list;
@@ -19,37 +20,36 @@ use crate::helpers::term::print_error;
 use clap::App;
 
 fn main() {
-    let matches = App::new("Bld")
+    let commands = vec![
+        auth::AuthCommand::boxed(),
+        config::ConfigCommand::boxed(), 
+        hist::HistCommand::boxed(),
+        init::InitCommand::boxed(),
+        inspect::InspectCommand::boxed(),
+        list::ListCommand::boxed(),
+        monit::MonitCommand::boxed(),
+        push::PushCommand::boxed(),
+        run::RunCommand::boxed(),
+        server::ServerCommand::boxed(),
+        stop::StopCommand::boxed(),
+    ];
+
+    let cli = App::new("Bld")
         .version(VERSION)
         .about("A simple CI/CD")
-        .subcommands(vec![
-            auth::command(),
-            init::command(),
-            inspect::command(),
-            hist::command(),
-            config::command(),
-            run::command(),
-            server::command(),
-            monit::command(),
-            list::command(),
-            push::command(),
-            stop::command(),
-        ])
+        .subcommands(commands
+            .iter()
+            .map(|c| c.interface())
+            .collect::<Vec<App<'static, 'static>>>())
         .get_matches();
 
-    let result = match matches.subcommand() {
-        ("login", Some(matches)) => auth::exec(matches),
-        ("init", Some(matches)) => init::exec(matches),
-        ("inspect", Some(matches)) => inspect::exec(matches),
-        ("hist", Some(matches)) => hist::exec(matches),
-        ("config", Some(matches)) => config::exec(matches),
-        ("run", Some(matches)) => run::exec(matches),
-        ("server", Some(matches)) => server::exec(matches),
-        ("monit", Some(matches)) => monit::exec(matches),
-        ("ls", Some(matches)) => list::exec(matches),
-        ("push", Some(matches)) => push::exec(matches),
-        ("stop", Some(matches)) => stop::exec(matches),
-        _ => Ok(()),
+    let result = match cli.subcommand() {
+        (id, Some(matches)) => commands
+            .iter()
+            .find(|c| c.id() == id)
+            .map(|c| c.exec(matches))
+            .unwrap_or_else(|| Ok(())),
+        _ => Ok(())
     };
 
     if let Err(e) = result {
