--- src/push/cli.rs
@@ -1,19 +1,0 @@
-use crate::config::definitions::VERSION;
-use clap::{App, Arg, SubCommand};
-
-pub fn command() -> App<'static, 'static> {
-    let pipeline = Arg::with_name("pipeline")
-        .short("p")
-        .long("pipeline")
-        .help("The name of the pipeline to push")
-        .takes_value(true);
-    let server = Arg::with_name("server")
-        .short("s")
-        .long("server")
-        .help("The name of the server to push changes to")
-        .takes_value(true);
-    SubCommand::with_name("push")
-        .about("Pushes the contents of a pipeline to a bld server")
-        .version(VERSION)
-        .args(&[pipeline, server])
-}
