--- src/run/cli.rs
@@ -1,29 +1,0 @@
-use crate::config::definitions::VERSION;
-use clap::{App, Arg, SubCommand};
-
-pub fn command() -> App<'static, 'static> {
-    let pipeline = Arg::with_name("pipeline")
-        .short("p")
-        .long("pipeline")
-        .help("Path to pipeline script")
-        .takes_value(true);
-    let server = Arg::with_name("server")
-        .short("s")
-        .long("server")
-        .help("The name of the server to run the pipeline")
-        .takes_value(true);
-    let detach = Arg::with_name("detach")
-        .short("d")
-        .long("detach")
-        .help("Detaches from the run execution (for server mode runs)");
-    let variables = Arg::with_name("variables")
-        .short("v")
-        .long("variables")
-        .help("Define values for variables of a pipeline")
-        .multiple(true)
-        .takes_value(true);
-    SubCommand::with_name("run")
-        .about("Executes a build pipeline")
-        .version(VERSION)
-        .args(&[pipeline, server, detach, variables])
-}
