+++ src/server/mod.rs
@@ -1,13 +1,11 @@
-mod cli;
+mod command;
 mod endpoints;
-mod exec;
 mod extractors;
 mod sockets;
 mod state;
 
-pub use cli::*;
+pub use command::*;
 pub use endpoints::*;
-pub use exec::*;
 pub use extractors::*;
 pub use sockets::*;
 pub use state::*;
