+++ src/types/command.rs
@@ -1,0 +1,10 @@
+use crate::types::Result;
+use clap::{App, ArgMatches};
+
+pub trait BldCommand {
+    fn id(&self) -> &'static str;
+
+    fn interface(&self) -> App<'static, 'static>;
+
+    fn exec(&self, matches: &ArgMatches<'_>) -> Result<()>;
+}
