+++ src/types/mod.rs
@@ -1,4 +1,5 @@
 mod auth;
+mod command;
 mod exec;
 mod monit;
 mod push;
@@ -6,6 +7,7 @@ mod recv;
 mod result;
 
 pub use auth::*;
+pub use command::*;
 pub use exec::*;
 pub use monit::*;
 pub use push::*;
